% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sc_DE_data.r
\name{format_data}
\alias{format_data}
\title{Formats DE from single cell into a matrix for heatmap}
\usage{
format_data(
  folder = NULL,
  DElist = NULL,
  filetype = ".txt",
  filecolsep = "\\t"
)
}
\arguments{
\item{folder}{folder with text files with significant DE genes for single cell}

\item{DElist}{list object with significant DE genes for single}

\item{filetype}{file type in folder with significant DE genes}

\item{filecolsep}{separator for columns in files with significant DE genes}
}
\description{
Formats DE from single cell into a matrix for heatmap
}
\examples{
data <- format_data(folder='path/to/significant/DEresults/')

}
\keyword{data}
\keyword{format}
\keyword{single-cell}
