% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_full_DE_heatmap.r
\name{sc_heatmap}
\alias{sc_heatmap}
\title{generates single cell heatmap}
\usage{
sc_heatmap(
  data,
  colorder = NULL,
  genes2remove = NULL,
  filename = "heatmap.png",
  cluster = TRUE,
  clusterbar = TRUE,
  kmeanscluster = TRUE,
  kclusters = 6,
  distmethod = "euclidean",
  clustermethod = "ward.D2",
  cuttreeparam = 6,
  figheight = 3,
  figwidth = 8,
  genefontsize = 10,
  legendfont = 6,
  labelfont = 6,
  labels = NULL,
  maxlfc = NULL,
  minlfc = NULL,
  linewidth = 4
)
}
\arguments{
\item{data}{data$data object from format_data function, to be run before this}

\item{colorder}{order that you want comparisons to show up on the heatmap}

\item{filename}{filename for heatmap (has to be .png, code will automatically generate pdf and svg versions of the heatmap)}

\item{cluster}{whether or not cluster genes in the heatmap}

\item{clusterbar}{whether to show clusters with a bar at the bottom of the heatmap}

\item{kmeanscluster}{whether to use kmeans clustering}

\item{kclusters}{Number of kmeans clusters if kmeanscluster=TRUE}

\item{distmethod}{distance method to use if kmeanscluster=FALSE (default=euclidean)}

\item{clustermethod}{cluster method to use default is set to ward.D2 but can uses any hclust method ("ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC))}

\item{cuttreeparam}{where to cut tree to generate clusters}

\item{figheight}{figure size height (inches)}

\item{figwidth}{figure size width (inches)}

\item{genefontsize}{size of gene names if clusterbar=FALSE}

\item{legendfont}{size of legend font}

\item{labelfont}{size of label (y axis) font}

\item{labels}{vector of specific labels}

\item{maxlfc}{maximum saturation level of LFC to show on heatmap}

\item{minlfc}{minimum saturation level of LFC to show on heatmap}

\item{linewidth}{maximum line width of individual genes markers in heatmap}
}
\description{
generates single cell heatmap
}
\examples{
clusters <- sc_heatmap(data$data)
}
\keyword{data}
\keyword{format}
\keyword{single-cell}
